refactor(utils): Group utils classes in utils folder
